se realiza mejoras de venatanas unicas al hacer click

+++ src/main/java/com/bbva/gateway/dto/iso20022/CardholderBillingAmountDTO.java
@@ -4,6 +4,7 @@ import lombok.*;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Builder
 @Getter
@@ -15,5 +16,6 @@ public class CardholderBillingAmountDTO implements Serializable {
 	private static final long serialVersionUID = 3499965191409678444L;
 	private Double amount;
 	private String currency;
-	private Double effectiveExchangeRate;
+	//private Double effectiveExchangeRate;
+	private BigDecimal effectiveExchangeRate;
 }
