+++ src/main/java/com/bbva/gateway/dto/iso20022/ReconciliationAmountDTO.java
@@ -4,6 +4,7 @@ import lombok.*;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 
 @Builder
@@ -16,6 +17,7 @@ public class ReconciliationAmountDTO implements Serializable {
 	private static final long serialVersionUID = -6264859555050322634L;
 	private Double amount;
 	private String currency;
-	private Double effectiveExchangeRate;
+	//private Double effectiveExchangeRate;
+	private BigDecimal effectiveExchangeRate;
 	private String conversionDate;
 }
